% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noa_get_tables.R
\name{noa_get_latest}
\alias{noa_get_latest}
\title{Get latest observations from weather stations}
\usage{
noa_get_latest()
}
\value{
a tibble
}
\description{
Get latest observations from weather stations
}
\examples{

library(tibble)
noa_get_latest()
}
