% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noar_package.R
\docType{package}
\name{noar}
\alias{noar}
\alias{_PACKAGE}
\alias{noar-package}
\alias{noar-package}
\title{noar: Interface to National Observatory of Athens weather stations network}
\description{
R interface to the National Observatory of Athens weather stations data
}
\section{Data Sources}{


The data are retrieved from http://www.meteo.gr/
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kvantas/noar}
  \item Report bugs at \url{https://github.com/kvantas/noar/issues}
}

}
\author{
\strong{Maintainer}: Konstantinos Vantas \email{kon.vantas@gmail.com} (0000-0001-6387-8791)

}
