% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noar_package.R
\docType{package}
\name{noar}
\alias{noar}
\alias{_PACKAGE}
\alias{noar-package}
\alias{noar-package}
\title{noar: Interface to National Observatory of Athens weather stations network}
\description{
R interface to the National Observatory of Athens weather stations
network http://www.meteo.gr/.
}
\section{Data Sources}{


The data are retrieved from http://www.meteo.gr/
}

\author{
\strong{Maintainer}: Konstantinos Vantas \email{kon.vantas@gmail.com} (0000-0001-6387-8791)

}
